Refactor CompressedTar into CompressedStreamArchive
This lets us reuse CompressedStreamArchive for other formats (e.g. rpm)
