+++ NanaZip.Core/Extensions/ZSCodecs/BrotliHandler.cpp
@@ -46,7 +46,7 @@ Z7_CLASS_IMP_CHandler_IInArchive_3(
 };
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
