+++ NanaZip.Core/Extensions/ZSCodecs/LzHandler.cpp
@@ -154,7 +154,7 @@ Z7_CLASS_IMP_CHandler_IInArchive_1(
 };
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
