+++ NanaZip.Core/SevenZip/CPP/7zip/Archive/Bz2Handler.cpp
@@ -47,7 +47,7 @@ Z7_CLASS_IMP_CHandler_IInArchive_3(
 };
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
