+++ NanaZip.Core/SevenZip/CPP/7zip/Archive/GzHandler.cpp
@@ -485,7 +485,7 @@ public:
 };
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
