+++ NanaZip.Core/SevenZip/CPP/7zip/Archive/Tar/TarHandler.cpp
@@ -30,7 +30,7 @@ namespace NTar {
 
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
