+++ NanaZip.Core/SevenZip/CPP/7zip/Archive/XzHandler.cpp
@@ -178,7 +178,7 @@ public:
 };
 
 // **************** NanaZip Modification Start ****************
-IInArchive *CreateArcForTar()
+IInArchive *CreateArcExported()
 {
     return new CHandler();
 }
